[v0.87][maintainability] Split runtime_artifacts execute tests by scenario
