+++ adl/tests/execute_tests/runtime_artifacts/artifact_failures.rs
@@ -1,0 +1,308 @@
+use super::*;
+
+#[test]
+fn run_reports_error_when_materialized_doc_is_missing() {
+    let base = tmp_dir("exec-missing-doc");
+    let _bin = write_mock_ollama(&base, MockOllamaBehavior::Success);
+    let new_path = prepend_path(&base);
+    let _path_guard = EnvVarGuard::set("PATH", new_path);
+
+    let names = ["doc_1.txt", "doc_2.txt", "doc_3.txt"];
+
+    fs::create_dir_all(base.join("examples/docs")).unwrap();
+    fs::create_dir_all(base.join("docs")).unwrap();
+
+    for name in names {
+        let src = Path::new("examples/docs").join(name);
+        fs::copy(&src, base.join("examples/docs").join(name)).unwrap();
+        fs::copy(&src, base.join("docs").join(name)).unwrap();
+    }
+
+    let yaml_src = fs::read_to_string("examples/adl-0.1.yaml").unwrap();
+    let yaml_broken = yaml_src
+        .replace(
+            "@file:examples/docs/doc_1.txt",
+            "@file:examples/docs/DOES_NOT_EXIST.txt",
+        )
+        .replace("@file:docs/doc_1.txt", "@file:docs/DOES_NOT_EXIST.txt");
+
+    let tmp_yaml = base.join("adl-broken.yaml");
+    fs::write(&tmp_yaml, yaml_broken.as_bytes()).unwrap();
+
+    let out = run_swarm(&[tmp_yaml.to_string_lossy().as_ref(), "--run"]);
+    assert!(
+        !out.status.success(),
+        "expected failure, got success.\nstdout:\n{}\nstderr:\n{}",
+        String::from_utf8_lossy(&out.stdout),
+        String::from_utf8_lossy(&out.stderr)
+    );
+
+    let stderr = String::from_utf8_lossy(&out.stderr);
+    assert!(
+        stderr.contains("failed to materialize inputs")
+            || stderr.contains("failed to stat input file"),
+        "stderr was:\n{stderr}"
+    );
+    assert!(stderr.contains("doc_1"), "stderr was:\n{stderr}");
+}
+
+#[test]
+fn run_surfaces_provider_failure_stderr() {
+    let base = tmp_dir("exec-provider-failure");
+    let _bin = write_mock_ollama(&base, MockOllamaBehavior::Fail);
+    let new_path = prepend_path(&base);
+    let _path_guard = EnvVarGuard::set("PATH", new_path);
+
+    let out = run_swarm(&["examples/adl-0.1.yaml", "--run"]);
+    assert!(
+        !out.status.success(),
+        "expected failure, got success.\nstdout:\n{}\nstderr:\n{}",
+        String::from_utf8_lossy(&out.stdout),
+        String::from_utf8_lossy(&out.stderr)
+    );
+
+    let stderr = String::from_utf8_lossy(&out.stderr);
+    assert!(
+        stderr.contains("ollama run failed") || stderr.contains("mock ollama failure"),
+        "stderr was:\n{stderr}"
+    );
+}
+
+#[test]
+fn run_writes_step_output_to_file() {
+    let base = tmp_dir("exec-write-output");
+    let _bin = write_mock_ollama(&base, MockOllamaBehavior::Success);
+    let new_path = prepend_path(&base);
+    let _path_guard = EnvVarGuard::set("PATH", new_path);
+
+    let yaml = r#"
+version: "0.2"
+
+providers:
+  local:
+    type: "ollama"
+    config:
+      model: "phi4-mini"
+
+agents:
+  a1:
+    provider: "local"
+    model: "phi4-mini"
+
+tasks:
+  t1:
+    prompt:
+      user: "Summarize: {{text}}"
+
+run:
+  name: "write-output"
+  workflow:
+    kind: "sequential"
+    steps:
+      - id: "s1"
+        agent: "a1"
+        task: "t1"
+        inputs:
+          text: "hello"
+        save_as: "summary"
+        write_to: "index.html"
+"#;
+
+    let tmp_yaml = base.join("write-output.yaml");
+    fs::write(&tmp_yaml, yaml.as_bytes()).unwrap();
+
+    let out_dir = base.join("out");
+    let out = run_swarm(&[
+        tmp_yaml.to_string_lossy().as_ref(),
+        "--run",
+        "--out",
+        out_dir.to_string_lossy().as_ref(),
+    ]);
+    assert!(
+        out.status.success(),
+        "expected success, got {:?}\nstderr:\n{}",
+        out.status.code(),
+        String::from_utf8_lossy(&out.stderr)
+    );
+
+    let output_path = out_dir.join("index.html");
+    let contents = fs::read_to_string(&output_path).unwrap();
+    assert!(
+        contents.contains("mock summary bullet one"),
+        "output file missing expected content: {contents}"
+    );
+
+    let stdout = String::from_utf8_lossy(&out.stdout);
+    assert!(
+        stdout.contains("ARTIFACT step=") && stdout.contains("index.html"),
+        "stdout was:\n{stdout}"
+    );
+}
+
+#[test]
+fn run_rejects_write_to_traversal() {
+    let base = tmp_dir("exec-write-traversal");
+    let _bin = write_mock_ollama(&base, MockOllamaBehavior::Success);
+    let new_path = prepend_path(&base);
+    let _path_guard = EnvVarGuard::set("PATH", new_path);
+
+    let yaml = r#"
+version: "0.2"
+
+providers:
+  local:
+    type: "ollama"
+    config:
+      model: "phi4-mini"
+
+agents:
+  a1:
+    provider: "local"
+    model: "phi4-mini"
+
+tasks:
+  t1:
+    prompt:
+      user: "Summarize: {{text}}"
+
+run:
+  name: "write-traversal"
+  workflow:
+    kind: "sequential"
+    steps:
+      - id: "bad-step"
+        agent: "a1"
+        task: "t1"
+        inputs:
+          text: "hello"
+        save_as: "summary"
+        write_to: "../escape.html"
+"#;
+
+    let tmp_yaml = base.join("write-traversal.yaml");
+    fs::write(&tmp_yaml, yaml.as_bytes()).unwrap();
+
+    let out = run_swarm(&[tmp_yaml.to_string_lossy().as_ref(), "--run"]);
+    assert!(
+        !out.status.success(),
+        "expected failure, got success.\nstdout:\n{}\nstderr:\n{}",
+        String::from_utf8_lossy(&out.stdout),
+        String::from_utf8_lossy(&out.stderr)
+    );
+
+    let stderr = String::from_utf8_lossy(&out.stderr);
+    assert!(
+        stderr.contains("bad-step") && stderr.contains("write_to"),
+        "stderr was:\n{stderr}"
+    );
+}
+
+#[test]
+fn run_rejects_missing_prompt_inputs() {
+    let base = tmp_dir("exec-missing-inputs");
+    let _bin = write_mock_ollama(&base, MockOllamaBehavior::Success);
+    let new_path = prepend_path(&base);
+    let _path_guard = EnvVarGuard::set("PATH", new_path);
+
+    let yaml = r#"
+version: "0.2"
+
+providers:
+  local:
+    type: "ollama"
+
+agents:
+  a1:
+    provider: "local"
+    model: "phi4-mini"
+
+tasks:
+  t1:
+    prompt:
+      user: "Summarize {{missing_key}}"
+
+run:
+  name: "missing-inputs"
+  workflow:
+    kind: "sequential"
+    steps:
+      - id: "s1"
+        agent: "a1"
+        task: "t1"
+        inputs:
+          text: "hello"
+"#;
+
+    let tmp_yaml = base.join("missing-inputs.yaml");
+    fs::write(&tmp_yaml, yaml.as_bytes()).unwrap();
+
+    let out = run_swarm(&[tmp_yaml.to_string_lossy().as_ref(), "--run"]);
+    assert!(
+        !out.status.success(),
+        "expected failure, got success.\nstdout:\n{}\nstderr:\n{}",
+        String::from_utf8_lossy(&out.stdout),
+        String::from_utf8_lossy(&out.stderr)
+    );
+
+    let stderr = String::from_utf8_lossy(&out.stderr);
+    assert!(
+        stderr.contains("missing input bindings") && stderr.contains("missing_key"),
+        "stderr was:\n{stderr}"
+    );
+}
+
+#[test]
+fn run_allows_prompt_only_step_with_no_inputs() {
+    let base = tmp_dir("exec-prompt-only-no-inputs");
+    let _bin = write_mock_ollama(&base, MockOllamaBehavior::Success);
+    let new_path = prepend_path(&base);
+    let _path_guard = EnvVarGuard::set("PATH", new_path);
+
+    let yaml = r#"
+version: "0.2"
+
+providers:
+  local:
+    type: "ollama"
+    config:
+      model: "phi4-mini"
+
+agents:
+  a1:
+    provider: "local"
+    model: "phi4-mini"
+
+tasks:
+  t1:
+    prompt:
+      user: "Summarize this prompt-only step."
+
+run:
+  name: "prompt-only-step"
+  workflow:
+    kind: "sequential"
+    steps:
+      - id: "s1"
+        agent: "a1"
+        task: "t1"
+"#;
+
+    let tmp_yaml = base.join("prompt-only.yaml");
+    fs::write(&tmp_yaml, yaml.as_bytes()).unwrap();
+
+    let out = run_swarm(&[tmp_yaml.to_string_lossy().as_ref(), "--run"]);
+    assert!(
+        out.status.success(),
+        "expected success for prompt-only step, got {:?}\nstdout:\n{}\nstderr:\n{}",
+        out.status.code(),
+        String::from_utf8_lossy(&out.stdout),
+        String::from_utf8_lossy(&out.stderr)
+    );
+
+    let stdout = String::from_utf8_lossy(&out.stdout);
+    assert!(stdout.contains("--- step: s1 ---"), "stdout was:\n{stdout}");
+    assert!(
+        stdout.contains("mock summary bullet one"),
+        "stdout was:\n{stdout}"
+    );
+}
